** Please see a UML chart of the ClusterItem logic. It is saved in Drive. **
Rebuilt the Partner and Plane classes so that they now have a common ClusterMarker class which they inherit.

The ClusterMarker is now a ClusterItem which is heavily modified.

The class MarkerRenderer specifies how Markers are drawn in a ClusterManager. The MarkerRenderer contains a HashMap<String, BitmapDescriptor>
which contains all "Images" that are used for drawing the Marker Icons. A specific method, setIcon() is used to fill the HashMap with
partnerName - BitmapDescriptor combinations that the MarkerRenderer can read when drawing Markers that emerge from inside a Cluster.

PartnerMarkerClass takes care of Partners and downloads them from Firebase.

PlaneMarkerClass takes care of Planes and downloads them from OpenSky.

Currently PlaneMarkerClass and PartnerMarkerClass are so un-similar that they do not require a parent class.